feat: add support for minimal install (i.e., without ros dependency)
- Move ros related dependency to the optional group in `pyproject.toml`
- Added conditional namespace logic for when ros dependencies are not available.
- Refactored imports in relevant modules to use a `try-except` structure,
  ensuring the error is raised consistently when ROS dependencies are missing.
- Added custom `RosImportError` to provide clearer feedback when ROS-related
  modules cannot be imported, enhancing user guidance during setup.
- Updated tests to verify proper handling and messaging of missing modules.

Issue TCT-73
