+++ .dockerized_norlab/configuration/Dockerfile
@@ -16,11 +16,10 @@ WORKDIR ${DN_PROJECT_PATH:?'environment variable is not set'}
 # ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓
 # ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓ ↓
 
-
 COPY  . "${DN_PROJECT_PATH:?err}/"
 RUN <<EOF
   n2st::print_msg "Install trajectory-container-tools via pyproject.toml..."
   cd "${DN_PROJECT_PATH}"
-  python -m pip install --editable .[dev] \
+  python -m pip install --editable .[ros,dev] \
     || n2st::print_msg_error_and_exit "Failed trajectory-container-tools install via pyproject.toml!"
 EOF
