+++ .dockerized_norlab/configuration/project_requirements/python.requirements-dna.txt
@@ -30,7 +30,7 @@ pyclean
 #setuptools-scm
 
 # .................................................................................................
-# (☕minor) ToDo: on task NMO-802 end >> validate it's still required ↓↓
+# (☕minor) ToDo: on task NMO-802 end >> validate it's not required anymore ↓↓
 # Note: Set DNA to build for Numpy 2 but let the pyproject.toml dependencies open so that we keep
 # support for Numpy 1
 #matplotlib >= 3.8.0  # These versions support NumPy 2.x
