+++ src/trajectory_container_tools/dataclasses/core/__init__.py
@@ -1,7 +1,10 @@
 # coding=utf-8
-from .abstract_multifeature_dataclass import AbstractMultifeatureDataclass, AbstractMultifeatureStampedDataclass
-from .abstract_no_trajectory_dataclass import AbstractNoTrajectoryDataclass
 from .abstract_trajectory_dataclass import AbstractTrajectoryDataclass
+from .abstract_no_trajectory_dataclass import AbstractNoTrajectoryDataclass
+from .abstract_multifeature_dataclass import AbstractMultifeatureDataclass
+from .abstract_multifeature_stamped_dataclass import (
+    AbstractMultifeatureStampedDataclass,
+)
 from .base_trajectory_dataclass import (
     BaseTrajectoryDataclass,
     BaseNoTrajectoryDataclass,
