+++ src/trajectory_container_tools/utils/ros2_utils/ros2_timestamps.py
@@ -1,34 +1,13 @@
 # coding=utf-8
-from dataclasses import dataclass
-from typing import Optional, Union
+from typing import Union
 
-from rclpy.time import Time as RosTime
-from rosbags.typesys.stores.empty import builtin_interfaces__msg__Time
+from trajectory_container_tools.utils.general import RosImportError
 
-
-@dataclass
-class TrajectoryTimestampsMetadata:
-    """Represents a trajectory timestamps metadata.
-
-    This class encapsulates the start time, end time, and duration of a trajectory. The `duration`
-    is automatically calculated if not provided during initialization. It can be used for various
-    time-based operations related to trajectory handling.
-
-    :ivar start_time: The start time of the trajectory.
-    :type start_time: int
-    :ivar end_time: The end time of the trajectory.
-    :type end_time: int
-    :ivar duration: The total duration of the trajectory.
-    :type duration: Optional[int]
-    """
-
-    start_time: int
-    end_time: int
-    duration: Optional[int] = None
-
-    def __post_init__(self):
-        if self.duration is None:
-            self.duration = self.end_time - self.start_time
+try:
+    from rclpy.time import Time as RosTime
+    from rosbags.typesys.stores.empty import builtin_interfaces__msg__Time
+except (ImportError, ModuleNotFoundError):
+    raise RosImportError
 
 
 def convert_timestamp_from_rosbag_message(
