+++ tests/tests_extractor/tests_dataframe/test_dataframe_to_trajectory_dataclass.py
