+++ tests/tests_extractor/tests_rosbag/test_rosbag_to_trajectory_dataclass.py
@@ -4,7 +4,7 @@ from typing import Union
 import pytest
 import numpy as np
 
-from ..rosbag_test_utils import get_rosbag_vaul_f110_grand_salon_path
+from tests.rosbag_test_utils import get_rosbag_vaul_f110_grand_salon_path
 from trajectory_container_tools.extractor.rosbag_to_tct import (
     from_rosbag,
     extract_rosbag_feature,
